Issue in the harvest function partly fixed
In the harvest function, there wasn't a case where the only card in the field is harvested, as it doesn't necessarily come under the case of number of cards for one coin, which was the last possible case before the fix. Now the case has been added and the harvest function works for every bean except garden bean.